+++ tianshou/policy/modelfree/bdqn.py
@@ -109,7 +109,7 @@ class BDQN(QLearningOffPolicyAlgorithm[BDQNPolicy]):
         gamma: float = 0.99,
         estimation_step: int = 1,
         target_update_freq: int = 0,
-        reward_normalization: bool = False,
+        return_scaling: bool = False,
         is_double: bool = True,
     ) -> None:
         """
@@ -131,8 +131,9 @@ class BDQN(QLearningOffPolicyAlgorithm[BDQNPolicy]):
             complete episode returns.
         :param target_update_freq: the target network update frequency (0 if
             you do not use the target network).
-        :param reward_normalization: normalize the **returns** to Normal(0, 1).
-            TODO: rename to return_normalization?
+        :param return_scaling: flag indicating whether to scale/standardise returns to Normal(0, 1) based
+            on running mean and standard deviation.
+            Support for this is currently suspended and therefore the flag should not be enabled.
         :param is_double: whether to use double DQN.
         """
         assert (
@@ -144,7 +145,6 @@ class BDQN(QLearningOffPolicyAlgorithm[BDQNPolicy]):
             gamma=gamma,
             estimation_step=estimation_step,
             target_update_freq=target_update_freq,
-            reward_normalization=reward_normalization,
         )
         self.is_double = is_double